+++ piecrust-uplink/src/error.rs
@@ -4,10 +4,15 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
+extern crate alloc;
+
 use bytecheck::CheckBytes;
 use rkyv::{Archive, Deserialize, Serialize};
 
+use alloc::string::String;
+
 use core::fmt::{Display, Formatter};
+use core::str;
 
 /// The error possibly returned on an inter-contract-call.
 //
@@ -17,30 +22,66 @@ use core::fmt::{Display, Formatter};
 //
 // The contract writer, however, is free to pass it around and react to it if it
 // wishes.
-#[derive(Debug, Clone, Copy, Archive, Serialize, Deserialize)]
+#[derive(Debug, Clone, Archive, Serialize, Deserialize)]
 #[archive_attr(derive(CheckBytes))]
 pub enum ContractError {
-    PANIC,
-    OUTOFGAS,
-    OTHER(i32),
+    Panic(String),
+    OutOfPoints,
+    Unknown,
 }
 
 impl ContractError {
-    /// Returns a contract error from a return `code`.
-    ///
-    /// # Panic
-    /// Panics if the value is larger than or equal to 0.
-    pub fn from_code(code: i32) -> Self {
-        if code >= 0 {
-            panic!(
-                "A `ContractError` is never equal or larger than 0, got {code}"
-            );
+    /// Returns a contract error from a return `code` and the data in the
+    /// `slice`.
+    #[cfg(feature = "abi")]
+    pub(crate) fn from_parts(code: i32, slice: &[u8]) -> Self {
+        fn get_msg(slice: &[u8]) -> String {
+            let msg_len = {
+                let mut msg_len_bytes = [0u8; 4];
+                msg_len_bytes.copy_from_slice(&slice[..4]);
+                u32::from_le_bytes(msg_len_bytes)
+            } as usize;
+
+            // SAFETY: the host guarantees that the message is valid UTF-8,
+            // so this is safe.
+            let msg = unsafe {
+                use alloc::string::ToString;
+                let msg_bytes = &slice[4..4 + msg_len];
+                let msg_str = str::from_utf8_unchecked(msg_bytes);
+                msg_str.to_string()
+            };
+
+            msg
         }
 
         match code {
-            -1 => Self::PANIC,
-            -2 => Self::OUTOFGAS,
-            v => Self::OTHER(v),
+            -1 => Self::Panic(get_msg(slice)),
+            -2 => Self::OutOfPoints,
+            i32::MIN => Self::Unknown,
+            _ => unreachable!("The host must guarantee that the code is valid"),
+        }
+    }
+
+    /// Write the appropriate data the `arg_buf` and return the error code.
+    pub fn to_parts(&self, slice: &mut [u8]) -> i32 {
+        fn put_msg(msg: &str, slice: &mut [u8]) {
+            let msg_bytes = msg.as_bytes();
+            let msg_len = msg_bytes.len();
+
+            let mut msg_len_bytes = [0u8; 4];
+            msg_len_bytes.copy_from_slice(&(msg_len as u32).to_le_bytes());
+
+            slice[..4].copy_from_slice(&msg_len_bytes);
+            slice[4..4 + msg_len].copy_from_slice(msg_bytes);
+        }
+
+        match self {
+            Self::Panic(msg) => {
+                put_msg(msg, slice);
+                -1
+            }
+            Self::OutOfPoints => -2,
+            Self::Unknown => i32::MIN,
         }
     }
 }
@@ -48,9 +89,9 @@ impl ContractError {
 impl From<ContractError> for i32 {
     fn from(err: ContractError) -> Self {
         match err {
-            ContractError::PANIC => -1,
-            ContractError::OUTOFGAS => -2,
-            ContractError::OTHER(c) => c,
+            ContractError::Panic(_) => -1,
+            ContractError::OutOfPoints => -2,
+            ContractError::Unknown => i32::MIN,
         }
     }
 }
@@ -58,9 +99,9 @@ impl From<ContractError> for i32 {
 impl Display for ContractError {
     fn fmt(&self, f: &mut Formatter<'_>) -> core::fmt::Result {
         match self {
-            ContractError::PANIC => write!(f, "CONTRACT PANIC"),
-            ContractError::OUTOFGAS => write!(f, "OUT OF GAS"),
-            ContractError::OTHER(c) => write!(f, "OTHER: {c}"),
+            ContractError::Panic(msg) => write!(f, "Panic: {msg}"),
+            ContractError::OutOfPoints => write!(f, "OutOfPoints"),
+            ContractError::Unknown => write!(f, "Unknown"),
         }
     }
 }